More cleanup

+++ src/System/comp_cond.cc
@@ -44,7 +44,7 @@ namespace Loci {
 #ifdef DEBUG
     // sanity check, all vertices should be scheduled
     digraph::vertexSet allvertices ;
-    for(int i=0;i< dag_sched.size();++i) 
+    for(size_t i=0;i< dag_sched.size();++i) 
       allvertices += dag_sched[i] ;
     warn(allvertices != dag.get_all_vertices()) ;
 #endif
